Change the RequestHandle header append to use += instead of append for null byte
append() will ignore appending any null terminating bytes (event if it is the only
thing in the string being appended).  Using += with '\n' appends an 'int' value
and correctly appends the null byte into the string.  This is important for when
the curl_slist of headers is generated from m_request_headers buffer otherwise
all the headers run on until the last header.. oops.