Split Response object from Request
The main reason to do this is to be
able to introduce arbitrary timer
calls to the client, unforunately
one cannot give the client the CURL*
handle at arbitrary times while
the CURL* handle is still executing
within the curl multi handle.  This
move is slowly migrating towards
removing the client of the library
from ever having any ownership over
the CURL* handles.