Add Async Timesup feature
Asynchronous requests that have
very small timeout values can have
trouble establishing connections
especially if they use TLS, etc.

To attempt to solve this problem a
timesup feature is introduced that
is a two tier timeout system.  The
asynchronous request is given a shorter
timesup value that the application
is willing to wait for this period
of time for a response and then a
normal timeout value for establishing
connections.  The application is
notified as usual of a 'timesup'
when the shorter timer expires
but won't be notified of the longer
timeout value if that also expires.
If the connection is established
in the time between timesup and timeout
then the response is thrown away
but the connection is maintained
so the next request might succeed
since it no longer needs to connect.