Rename timesup to connection time feature
This now only exposes a timeout to the client
but allows for the client to add a connection time
value to each event loop.  This is a secondary
timeout value that the event loop will automatically
add to requests that have timeouts shorter than
the connection time value making the feature easier to use.