Move async callback or future to client
This removes the ability to set async on completion methods
directly from the request object.  I think this cleans up the
api a bit since those methods shouldn't be used on sync requests
and were only useful when using a request through an async client.

This now makes it more obvious that for any given requst through
a client the user must provide either a lambda to be called on
completion or to block/wait on a future.