Cleanup api, fix benchmark segfault
This change exposed a use after free in the
on complete handlers, in that it was being used
as a ref in the request copmleted function on client
but then immediately moved, this seemed to generally
work when it was just and std::function but now that
its a variant with some weird copy_but_actually_move
wrapper its clearly a use after free scenario.

Moving the wrapper object to be owned by the client
prior to invoking whichever 'on complete' method the
user wants resolves this lifetime issue.