async_promise_type possibly unitialized value
The std::variant for callback or future/promise with gcc-13 seems to be
triggering an uninitialized value warning. We can tell it isn't
uninitialized but the compiler seems to be thinking the 2nd type in the
variant, which is the callback type, is unitialized _when_ the promise
type, 3rd type in the variant, is the one being initialized.

Closes #153