Add lift::client_pool
This structure allows for the user to spin up N clients that will round
robin handle requests that are started.

Updated the benchmark.cpp file to use the new lift::client_pool instead
of spinning up N lift::client's into a vector. This test is _probably_
slower since requests will now bounce around to different threads, but I
think this is worth it with the ease of use of the lift::client_pool.

Closes #157