Remove lift share functionality (#170)
This feature has been broken in libcurl for as long as this project has
existed. As far as I know it is unlikely to ever be fixed, its cross
thread segfaults when it is used which is not acceptable, so I am
removing the feature from liblifthttp. If libcurl does ever fix this
feature then it can be re-added at that time.

Closes #165